+++ configuration.hpp
@@ -3,7 +3,9 @@
 #include "midi_control_change_mapping.hpp"
 #include "midi_event.hpp"
 
+#include <cstddef>
 #include <filesystem>
+#include <optional>
 #include <stdexcept>
 #include <string>
 #include <vector>
@@ -33,13 +35,15 @@ struct SoundFontDefinition {
     std::filesystem::path file;
     int bank{};
     int preset{};
+    std::optional<int> key;
 };
 
 struct ApplicationConfig {
     std::vector<SoundFontDefinition> soundfonts;
     std::vector<MidiControlChangeMapping> midi_control_change_mappings;
     MidiControlBinding recording_control;
-    MidiControlBinding next_soundfont_control;
+    std::optional<MidiControlBinding> next_soundfont_control;
+    std::optional<MidiControlBinding> soundfont_by_note_control;
     MidiControlBinding octave_down_control;
     MidiControlBinding octave_up_control;
 };